fix: AxmStatusReadActVel uses ref float (was ref double — type mismatch)

+++ Controller/AjinController.cs
@@ -47,7 +47,7 @@ namespace PHM_Project_DockPanel.Controller
         [DllImport(DLL, CallingConvention = CallingConvention.Cdecl)] private static extern uint AxmStatusSetActPos(int lAxisNo, double dActPos);
         [DllImport(DLL, CallingConvention = CallingConvention.Cdecl)] private static extern uint AxmStatusSetCmdPos(int lAxisNo, double dCmdPos);
         [DllImport(DLL, CallingConvention = CallingConvention.Cdecl)] private static extern uint AxmStatusGetCmdPos(int lAxisNo, ref double dpCmdPos);
-        [DllImport(DLL, CallingConvention = CallingConvention.Cdecl)] private static extern uint AxmStatusReadActVel(int lAxisNo, ref double dpActVel);
+        [DllImport(DLL, CallingConvention = CallingConvention.Cdecl)] private static extern uint AxmStatusReadActVel(int lAxisNo, ref float dpActVel);
         [DllImport(DLL, CallingConvention = CallingConvention.Cdecl)] private static extern uint AxmStatusReadTorque(int lAxisNo, ref double dpTorque);
 
         // ────────────────────────────────────────────────────────────
@@ -240,7 +240,7 @@ namespace PHM_Project_DockPanel.Controller
         public double GetActVel(int axisNo)
         {
             CheckConnected();
-            double vel = 0;
+            float vel = 0;
             uint ret = AxmStatusReadActVel(axisNo, ref vel);
             if (ret != AXT_RT_SUCCESS)
                 AppEvents.RaiseLog($"[Ajin] AxmStatusReadActVel axis={axisNo} 실패 (0x{ret:X})");
